[APP-0000] Fix RTL, fix imports
