+++ modules/scanner/assets/js/components/block-button/button-menu.js
@@ -9,8 +9,8 @@ import MenuItem from '@elementor/ui/MenuItem';
 import MenuItemIcon from '@elementor/ui/MenuItemIcon';
 import MenuItemText from '@elementor/ui/MenuItemText';
 import PropTypes from 'prop-types';
+import { BLOCK_TITLES } from '@ea11y-apps/global/constants';
 import { DeletePageRemediationModal } from '@ea11y-apps/scanner/components/remediation-confirmation-modal';
-import { BLOCK_TITLES } from '@ea11y-apps/scanner/constants';
 import { useScannerWizardContext } from '@ea11y-apps/scanner/context/scanner-wizard-context';
 import { useManageActions } from '@ea11y-apps/scanner/hooks/use-manage-actions';
 import { StyledBanIcon } from '@ea11y-apps/scanner/styles/app.styles';
