+++ modules/scanner/assets/js/components/main-list/index.js
@@ -1,5 +1,6 @@
+import { BLOCK_TITLES } from '@ea11y-apps/global/constants';
 import { BlockButton } from '@ea11y-apps/scanner/components/block-button';
-import { BLOCK_TITLES, BLOCKS } from '@ea11y-apps/scanner/constants';
+import { BLOCKS } from '@ea11y-apps/scanner/constants';
 import { useScannerWizardContext } from '@ea11y-apps/scanner/context/scanner-wizard-context';
 import { StyledBlockButtonsBox } from '@ea11y-apps/scanner/styles/app.styles';
