+++ package.json
@@ -48,7 +48,7 @@
 		"@elementor/design-tokens": "^1.1.4",
 		"@elementor/elementor-one-assets": "^0.4.16",
 		"@elementor/icons": "^1.46.0",
-		"@elementor/ui": "^1.36.14",
+		"@elementor/ui": "^1.37.3",
 		"@emotion/cache": "^11.14.0",
 		"@emotion/react": "^11.14.0",
 		"@mui/x-charts": "^7.27.0",
